libgnome-desktop-3-17.symbols: Update
The patch add two symbols and drops one. These symbols are defined in
"gnome-rr-private.h", which is not shipped in the -dev package, so they
can't be used in practice. This is a technical ABI break but one which
it's OK to overlook. There's probably a better way to handle the
visibility of these symbols, though.