debian: Move tests and debug utilities to new packages
Upstream is now installing more tests, and these are not something we
expect to have in a library package.

So, let's add two new packages for containing both the compiled and the
installed-tests metadata so that we can reduce the customizations.

As per this we can in fact use /usr/libexec for test programs and avoid
using a custom versioned path.

Gbp-Dch: Full