Rename source package to gnome-desktop
This is done now because we are adding new GTK4 packages
in addition to the GTK3 packages which are still built.

Eventually GNOME will stop offering a GTK3 build but
we may still have packages in Debian that need that library.
At that point, we'll need to reintroduce a gnome-desktop3
source package to keep things working.